-e JL ArgonV3 2024-08-18T09:27:31-04:00
Here is my progress report or end of the day report: I am implementing the AST, but first I am having to dive into the theory and design patterns that contribute to the function and design of the AST and the way it is incorporated. I am still a little confused with the visitor pattern and I need to continue studying 5.2.1 Disoriented obkjects -> 5.3.2 The visitor pattern. Additionally, i need to focus heavily on 5.3 Working with trees and 5.3.1 The expression problem.
 ---f